refactoring 'Task' to be HomePage
