+++ front/package.json
@@ -14,6 +14,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-modal": "^3.13.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "typescript": "^4.1.2",
     "web-vitals": "^1.0.1"
@@ -43,6 +44,7 @@
     ]
   },
   "devDependencies": {
-    "@types/react-modal": "^3.12.0"
+    "@types/react-modal": "^3.12.0",
+    "@types/react-router-dom": "^5.1.7"
   }
 }
