--- front/src/Componentes/tasks.css
@@ -1,105 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Lobster&display=swap");
-.titulo {
-  font-family: "Lobster", cursive;
-  font-style: normal;
-  font-weight: 400;
-  font-size: 72px;
-  color: #ffffff;
-}
-
-.t {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  height: 30vh;
-}
-
-.adicionar {
-  background-color: #26242d;
-  border: 1px solid #4e466a;
-  width: 90px;
-  height: 60px;
-  border-radius: 12%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  outline: none;
-  cursor: pointer;
-}
-
-.card {
-  display: grid;
-  grid-template-columns: repeat(5, 1fr);
-  font-family: "Lobster", cursive;
-  column-gap: 45px;
-  row-gap: 45px;
-}
-
-/* .a {
-  height: 160px;
-  width: 160px;
-  display: flex;
-  justify-content: flex-start;
-  align-items: flex-end;
-  padding: 0 0 20px 20px;
-  font-size: 20px;
-  border-radius: 10px;
-  box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.8);
-  background-color: #a5b9ff;
-} */
-/* .scroll{
-  overflow: scroll;
-} */
-.c {
-  height: 160px;
-  width: 160px;
-  display: flex;
-  justify-content: flex-start;
-  align-items: flex-end;
-  padding: 0 0 20px 20px;
-  font-size: 20px;
-  border-radius: 10px;
-  box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.8);
-  background-color: #a5b9ff;
-  word-break: break-all;
-}
-
-/* .c p {
-  font-size: 20px;
-  word-wrap: break-word;
-
-} */
-
-/* .b {
-  height: 160px;
-  width: 160px;
-  display: flex;
-  justify-content: flex-start;
-  align-items: flex-end;
-  padding: 0 0 20px 20px;
-  font-size: 20px;
-  border-radius: 10px;
-  box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.8);
-  background-color: #feb0ff;
-} */
-
-/* .c:nth-child(-n + 5) {
-  background-color: #a5b9ff;
-}
-
-.c:nth-child(n + 6) {
-  background-color: #feb0ff;
-}
-
-.c:nth-child(n + 11) {
-  background-color: #ffffe0;
-} */
-
-.task {
-  padding: 0px 100px 30px 100px;
-}
-
-body {
-  background-color: #26242d;
-}
